NPC/NpcEffect: use s32 idx + (u16) cast on use for matrix-pointer math
Target's source stores nameTab->getIndex() result as s32 (no truncation
at assignment), then casts to u16 just before mulli to extract the
node-matrix offset. Our u16 idx was forcing an immediate clrlwi after
the bl, shifting subsequent register allocation.

setNoteEffectMtxPtr_   86.14 -> 99.83% (+13.69pp; basically matching)
setHappyEffectMtxPtr_  86.14 -> 91.29% (+5.15pp)
setPollutionEffectMtxPtr_ 80.93 -> 93.31% (+12.38pp)
TU 76.95 -> 78.87 (+1.92pp); project 67.5875 -> 67.5897 (+0.0022pp).

+++ src/NPC/NpcEffect.cpp
@@ -50,10 +50,10 @@ void TBaseNPC::setHappyEffectMtxPtr_(const JUTNameTab* nameTab)
 		}
 	}
 	if (jntName) {
-		u16 idx       = nameTab->getIndex(jntName);
+		s32 idx       = nameTab->getIndex(jntName);
 		J3DModel* mdl = getModel();
 		mPtrHappyEffectMtx
-		    = (MtxPtr)((u8*)mdl->mNodeMatrices + idx * sizeof(Mtx));
+		    = (MtxPtr)((u8*)mdl->mNodeMatrices + (u16)idx * sizeof(Mtx));
 	}
 }
 
@@ -71,10 +71,10 @@ void TBaseNPC::setNoteEffectMtxPtr_(const JUTNameTab* nameTab)
 		break;
 	}
 	if (jntName) {
-		u16 idx       = nameTab->getIndex(jntName);
+		s32 idx       = nameTab->getIndex(jntName);
 		J3DModel* mdl = getModel();
 		mPtrNoteEffectMtx
-		    = (MtxPtr)((u8*)mdl->mNodeMatrices + idx * sizeof(Mtx));
+		    = (MtxPtr)((u8*)mdl->mNodeMatrices + (u16)idx * sizeof(Mtx));
 	}
 }
 
@@ -90,14 +90,14 @@ void TBaseNPC::setPollutionEffectMtxPtr_(const JUTNameTab* nameTab)
 	if (!isNormalMonteM() && !isNormalMonteW())
 		isMonte = false;
 	if (isMonte) {
-		u16 idxL       = nameTab->getIndex(sFootL);
+		s32 idxL       = nameTab->getIndex(sFootL);
 		J3DModel* mdlL = getModel();
 		mPtrPollutionLEffectMtx
-		    = (MtxPtr)((u8*)mdlL->mNodeMatrices + idxL * sizeof(Mtx));
-		u16 idxR       = nameTab->getIndex(sFootR);
+		    = (MtxPtr)((u8*)mdlL->mNodeMatrices + (u16)idxL * sizeof(Mtx));
+		s32 idxR       = nameTab->getIndex(sFootR);
 		J3DModel* mdlR = getModel();
 		mPtrPollutionREffectMtx
-		    = (MtxPtr)((u8*)mdlR->mNodeMatrices + idxR * sizeof(Mtx));
+		    = (MtxPtr)((u8*)mdlR->mNodeMatrices + (u16)idxR * sizeof(Mtx));
 		jntName = sKoshiNull;
 	} else {
 		bool isMare = true;
@@ -110,10 +110,10 @@ void TBaseNPC::setPollutionEffectMtxPtr_(const JUTNameTab* nameTab)
 		}
 	}
 	if (jntName) {
-		u16 idx       = nameTab->getIndex(jntName);
+		s32 idx       = nameTab->getIndex(jntName);
 		J3DModel* mdl = getModel();
 		mPtrPollutionEffectMtx
-		    = (MtxPtr)((u8*)mdl->mNodeMatrices + idx * sizeof(Mtx));
+		    = (MtxPtr)((u8*)mdl->mNodeMatrices + (u16)idx * sizeof(Mtx));
 	}
 }
